pyproject.toml
[project]
name = "eval-fw"
version = "0.1.0"
description = "AI Security Evaluation Framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.0",
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "rich>=14.0.0",
    "typer>=0.12.0",
    "weasyprint>=62.0",
]

[project.scripts]
eval-fw = "eval_fw.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eval_fw"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.black]
line-length = 100
target-version = ["py312"]

[dependency-groups]
dev = [
    "black>=24.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
